Rename: 페이지 별 컴포넌트 분리 및 CSR과 SSR 별도 적용
